Implement a minimal, experimental version of tail call optimizations for module instantiations.
This version will expand user module instantiations and if/else statements, and *nothing else*. Many more modules could be added to the TCO inliner, but this version is very minimal.

Assumes that nested group() invocations can safely be flattened.

This is an experiment / proof of concept. There is a bunch of quick and dirty copy/paste in here that a real implementation should do sensibly, and little attention has been paid to sensible structural refactoring. Use at your own risk.